DashboardSrv: add saveDashboard method
